+++ utils/include/rawrbox/utils/timer.hpp
@@ -10,9 +10,9 @@ namespace rawrbox {
 	protected:
 		std::string _id = "";
 
-		uint64_t _nextTick = 0;
-		uint64_t _msDelay = 0;
-		uint64_t _pausedTime = 0;
+		float _nextTick = 0;
+		float _msDelay = 0;
+		float _pausedTime = 0;
 
 		int _iterations = -1;
 		int _ticks = 0;
@@ -23,7 +23,7 @@ namespace rawrbox {
 		bool _paused = false;
 		bool _infinite = false;
 
-		static rawrbox::TIMER* init(const std::string& id, int reps, uint64_t msDelay, std::function<void()> func, std::function<void()> onComplete = nullptr);
+		static rawrbox::TIMER* init(const std::string& id, int reps, float msDelay, std::function<void()> func, std::function<void()> onComplete = nullptr);
 
 	public:
 		static uint32_t ID;
@@ -32,10 +32,10 @@ namespace rawrbox {
 		// STATIC ----
 		static void update();
 
-		static rawrbox::TIMER* simple(const std::string& id, uint64_t msDelay, std::function<void()> func, std::function<void()> onComplete = nullptr);
-		static rawrbox::TIMER* simple(uint64_t msDelay, std::function<void()> func, std::function<void()> onComplete = nullptr);
-		static rawrbox::TIMER* create(const std::string& id, int reps, uint64_t msDelay, std::function<void()> func, std::function<void()> onComplete = nullptr);
-		static rawrbox::TIMER* create(int reps, uint64_t msDelay, std::function<void()> func, std::function<void()> onComplete = nullptr);
+		static rawrbox::TIMER* simple(const std::string& id, float msDelay, std::function<void()> func, std::function<void()> onComplete = nullptr);
+		static rawrbox::TIMER* simple(float msDelay, std::function<void()> func, std::function<void()> onComplete = nullptr);
+		static rawrbox::TIMER* create(const std::string& id, int reps, float msDelay, std::function<void()> func, std::function<void()> onComplete = nullptr);
+		static rawrbox::TIMER* create(int reps, float msDelay, std::function<void()> func, std::function<void()> onComplete = nullptr);
 
 		static bool destroy(const std::string& id);
 		static bool pause(const std::string& id, bool pause);
